+++ crates/op-rbuilder/src/primitives/mod.rs
@@ -1,1 +1,4 @@
 pub mod reth;
+mod helpers;
+#[cfg(not(feature = "flashblocks"))]
+pub use crate::primitives::helpers::{estimate_gas_for_builder_tx, signed_builder_tx};
