+++ crates/op-rbuilder/src/primitives/reth/mod.rs
@@ -1,2 +1,6 @@
 mod execution;
 pub use execution::{ExecutedPayload, ExecutionInfo};
+mod payload_builder_ctx;
+pub use payload_builder_ctx::OpPayloadBuilderCtx;
+
+
